+++ blog/__init__.py
@@ -1,33 +1,38 @@
-import os
-
 from flask import Flask
 from flask_mail import Mail
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 from flask_sqlalchemy import SQLAlchemy
 
+from blog.config import Config
 
-app = Flask(__name__)
-app.config["SECRET_KEY"] = os.environ.get("FLASK_BLOG_SECRET")
 
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+db = SQLAlchemy()
 
-app.config["MAIL_SERVER"] = "smtp.googlemail.com"
-app.config["MAIL_PORT"] = 587
-app.config["MAIL_USE_TLS"] = True
-app.config["MAIL_USERNAME"] = os.environ.get("EMAIL_USERNAME")
-app.config["MAIL_PASSWORD"] = os.environ.get("EMAIL_PASSWORD")
+bcrypt = Bcrypt()
 
-db = SQLAlchemy(app)
-
-bcrypt = Bcrypt(app)
-
-login_manager = LoginManager(app)
-login_manager.login_view = "login"
+login_manager = LoginManager()
+login_manager.login_view = "users.login"
 login_manager.login_message_category = "info"
 
-mail = Mail(app)
+mail = Mail()
 
 
-from blog import routes
+def create_app(config_class=Config):
+	app = Flask(__name__)
+	app.config.from_object(config_class)
+
+	db.init_app(app)
+	bcrypt.init_app(app)
+	login_manager.init_app(app)
+	mail.init_app(app)
+
+	from blog.blog.routes import blog
+	from blog.posts.routes import posts
+	from blog.users.routes import users
+
+	app.register_blueprint(blog)
+	app.register_blueprint(posts)
+	app.register_blueprint(users)
+
+	return app
