+++ blog/config.py
@@ -1,0 +1,12 @@
+import os
+
+
+class Config:
+	SECRET_KEY = os.environ.get("FLASK_BLOG_SECRET")
+	SQLALCHEMY_DATABASE_URI = os.environ.get("FLASK_BLOG_DB_URI")
+	SQLALCHEMY_TRACK_MODIFICATIONS = False
+	MAIL_SERVER = "smtp.googlemail.com"
+	MAIL_PORT = 587
+	MAIL_USE_TLS = True
+	MAIL_USERNAME = os.environ.get("FLASK_BLOG_EMAIL_USERNAME")
+	MAIL_PASSWORD = os.environ.get("FLASK_BLOG_EMAIL_PASSWORD")
