+++ blog/templates/layout.html
@@ -27,7 +27,7 @@
 	  	<nav class="navbar navbar-expand-md navbar-dark bg-steel fixed-top">
 	    	<div class="container">
 	      		
-	      		<a class="navbar-brand mr-4" href="{{ url_for('index') }}">Flask Blog</a>
+	      		<a class="navbar-brand mr-4" href="{{ url_for('blog.index') }}">Flask Blog</a>
 		      	
 		      	<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle" aria-controls="navbarToggle" aria-expanded="false" aria-label="Toggle navigation">
 		        	<span class="navbar-toggler-icon"></span>
@@ -36,19 +36,19 @@
 			    <div class="collapse navbar-collapse" id="navbarToggle">
 		        
 		        	<div class="navbar-nav mr-auto">
-		          		<a class="nav-item nav-link" href="{{ url_for('index') }}">Home</a>
-		          		<a class="nav-item nav-link" href="{{ url_for('about') }}">About</a>
+		          		<a class="nav-item nav-link" href="{{ url_for('blog.index') }}">Home</a>
+		          		<a class="nav-item nav-link" href="{{ url_for('blog.about') }}">About</a>
 		        	</div>
 		        
 		        	<!-- Navbar Right Side -->
 		        	<div class="navbar-nav">
 		        		{% if current_user.is_authenticated %}
-		        		<a class="nav-item nav-link" href="{{ url_for('create_post') }}">New Post</a>
-					        <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-					        <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+		        		<a class="nav-item nav-link" href="{{ url_for('posts.create_post') }}">New Post</a>
+					        <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>
+					        <a class="nav-item nav-link" href="{{ url_for('users.logout') }}">Logout</a>
 				        {% else %}
-					        <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-					        <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+					        <a class="nav-item nav-link" href="{{ url_for('users.login') }}">Login</a>
+					        <a class="nav-item nav-link" href="{{ url_for('users.register') }}">Register</a>
 				        {% endif %}
 	        		</div>
