+++ blog/templates/post.html
@@ -7,12 +7,12 @@
 	<div class="media-body">
 		
 		<div class="article-metadata">
-			<a class="mr-2" href="{{ url_for('user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
+			<a class="mr-2" href="{{ url_for('posts.user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
 			<small class="text-muted">{{ post.date_posted.strftime("%d/%m/%Y") }}</small>
 
 			{% if post.author == current_user %}
 			<div>
-				<a class="btn btn-secondary btn-sm mt-1 mb-1" href="{{ url_for('update_post', post_id=post.id) }}">Update</a>
+				<a class="btn btn-secondary btn-sm mt-1 mb-1" href="{{ url_for('posts.update_post', post_id=post.id) }}">Update</a>
 				<button class="btn btn-danger btn-sm m-1" type="button" data-toggle="modal" data-target="#delete_post_modal">Delete</button>
 			</div>
 			{% endif %}
@@ -39,7 +39,7 @@
       </div>
       <div class="modal-footer">
         <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-        <form action="{{ url_for('delete_post', post_id=post.id) }}" method="POST">
+        <form action="{{ url_for('posts.delete_post', post_id=post.id) }}" method="POST">
         	<input type="submit" class="btn btn-danger" value="Delete">
         </form>
       </div>
