+++ blog/templates/register.html
@@ -80,7 +80,7 @@
 
 	<div class="border-top pt-3">
 		<small class="text-muted">
-			Already have an account? Login <a href="{{ url_for('login') }}">here</a>.
+			Already have an account? Login <a href="{{ url_for('users.login') }}">here</a>.
 		</small>
 	</div>
