+++ blog/templates/user_posts.html
@@ -9,10 +9,10 @@
 	<img class="rounded-circle article-img" src="{{ url_for('static', filename='account_icons/' + post.author.account_icon) }}">
   	<div class="media-body">
     	<div class="article-metadata">
-      		<a class="mr-2" href="{{ url_for('user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
+      		<a class="mr-2" href="{{ url_for('posts.user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
       		<small class="text-muted">{{ post.date_posted.strftime("%d/%m/%Y") }}</small>
     	</div>
-    	<h2><a class="article-title" href="{{ url_for('post', post_id=post.id) }}">{{ post.title }}</a></h2>
+    	<h2><a class="article-title" href="{{ url_for('posts.post', post_id=post.id) }}">{{ post.title }}</a></h2>
     	<p class="article-content">{{ post.content }}</p>
   	</div>
 </article>
@@ -21,9 +21,9 @@
 {% for page_num in posts.iter_pages(left_edge=2, right_edge=2, left_current=2, right_current=3) %}
 	{% if page_num %}
 		{% if posts.page == page_num %}
-		<a class="btn btn-info mb-4" href="{{ url_for('user_posts', username=user.username, page=page_num) }}">{{ page_num }}</a>
+		<a class="btn btn-info mb-4" href="{{ url_for('posts.user_posts', username=user.username, page=page_num) }}">{{ page_num }}</a>
 		{% else %}
-		<a class="btn btn-outline-info mb-4" href="{{ url_for('user_posts', username=user.username, page=page_num) }}">{{ page_num }}</a>
+		<a class="btn btn-outline-info mb-4" href="{{ url_for('posts.user_posts', username=user.username, page=page_num) }}">{{ page_num }}</a>
 		{% endif %}
 	{% else %}
