+++ blog/users/forms.py
@@ -1,9 +1,9 @@
-from blog.models import User
 from flask_wtf import FlaskForm
 from flask_wtf.file import FileField, FileAllowed
-from flask_login import current_user
-from wtforms import StringField, PasswordField, SubmitField, BooleanField, TextAreaField
+from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
+from flask_login import current_user
+from blog.models import User
 
 
 class RegisterForm(FlaskForm):
@@ -54,16 +54,6 @@ class UpdateAccountForm(FlaskForm):
 
 			if user:
 				raise ValidationError("Email exists. Please use a different one.")
-
-
-class CreatePostForm(FlaskForm):
-	title = StringField("Title", validators=[DataRequired(), Length(max=100)])
-	content = TextAreaField("Content", validators=[DataRequired()])
-	submit = SubmitField("Post")
-
-
-class UpdatePostForm(CreatePostForm):
-	pass
 
 
 class RequestResetPasswordForm(FlaskForm):
