+++ run.py
@@ -1,4 +1,7 @@
-from blog import app
+from blog import create_app
+
+
+app = create_app()
 
 
 if __name__ == "__main__":
